mavsdk_tests: Split tests and test main
